requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-groq==0.0.1
langchain-community==0.0.27
python-dotenv==1.0.1
faiss-cpu==1.7.4
unstructured==0.12.4
pdf2image==1.17.0
pytesseract==0.3.10
python-magic==0.4.27
python-magic-bin==0.4.14
pypdf==4.0.2
tiktoken==0.6.0